% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{check_query_status}
\alias{check_query_status}
\title{Report an error from API call}
\usage{
check_query_status(status)
}
\arguments{
\item{status}{The status from the API.}
}
\description{
Report an error from API call
}
\keyword{internal}
