% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords.R
\name{fetch_lords_addresses}
\alias{fetch_lords_addresses}
\title{Fetch addresses for all Lords}
\usage{
fetch_lords_addresses()
}
\value{
A tibble of addresses for each Lord, with one row per address.
}
\description{
\code{fetch_lords_addresses} fetches data from the Members Names platform
showing contact details for each Lord.
}
\details{
Addresses can represent contact information of different types, including
phsyical addresses, phone, fax, email, website, and social media. These
addresses are not time bound in MNIS so date filtering is not available for
this function.
}
