% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords.R
\name{fetch_lords_contested_elections}
\alias{fetch_lords_contested_elections}
\title{Fetch contested elections for all Lords}
\usage{
fetch_lords_contested_elections(from_date = NA, to_date = NA, on_date = NA)
}
\arguments{
\item{from_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the from_date.}

\item{to_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. 2000-12-31'. The
default value is NA, which means no records are excluded on the basis
of the to_date.}

\item{on_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. 2000-12-31'. The
default value is NA, which means no records are excluded on the basis
of the on_date.}
}
\value{
A tibble of contested elections for each Lord, with one row per
contested election.
}
\description{
\code{fetch_lords_contested_elections} fetches data from the Members Names
platform showing contested elections for each Lord.
}
\details{
The from_date and to_date arguments can be used to filter the contested
elections returned. The on_date argument is a convenience that sets the
from_date and to_date to the same given date. The on_date has priority:
if the on_date is set, the from_date and to_date are ignored.

The filtering is inclusive: a contested election is returned if
the date of it falls within the period specified with the from and to dates.
}
