% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_lord.R
\name{fetch_lords_contested_elections_raw}
\alias{fetch_lords_contested_elections_raw}
\title{Fetch contested elections: Lords}
\usage{
fetch_lords_contested_elections_raw()
}
\description{
Fetch contested elections: Lords
}
\keyword{internal}
