% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts_lords.R
\name{fetch_lords_facebook}
\alias{fetch_lords_facebook}
\title{Fetch Facebook accounts for Lords}
\usage{
fetch_lords_facebook(from_date = NA, to_date = NA, on_date = NA)
}
\arguments{
\item{from_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the from_date.}

\item{to_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the to_date.}

\item{on_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the on_date.}
}
\value{
A tibble of known Facebook accounts for each Lord, with one row per
  combination of Lord and Facebook account.
}
\description{
\code{fetch_lords_facebook} fetches data from the Members Names platform on
Facebook accounts for each Lord, with one row per combination of Lord and
Facebook account.
}
\details{
The from_date and to_date arguments can be used to filter the Lord based on
the dates of their Lords memberships. The on_date argument is a convenience
that sets the from_date and to_date to the same given date. The on_date has
priority: if the on_date is set, the from_date and to_date are ignored.

The filtering is inclusive: a Lord is returned if any part of one of their
Lords memberships falls within the period specified with the from and to
dates.
}
