% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords.R
\name{fetch_lords_maiden_speeches}
\alias{fetch_lords_maiden_speeches}
\title{Fetch maiden speeches for all Lords}
\usage{
fetch_lords_maiden_speeches(from_date = NA, to_date = NA, on_date = NA)
}
\arguments{
\item{from_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the from_date.}

\item{to_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. 2000-12-31'. The
default value is NA, which means no records are excluded on the basis
of the to_date.}

\item{on_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. 2000-12-31'. The
default value is NA, which means no records are excluded on the basis
of the on_date.}
}
\value{
A tibble of maiden speeches for each Lord, with one row per
  maiden speech.
}
\description{
\code{fetch_lords_maiden_speeches} fetches data from the Members Names platform
showing maiden speeches for each Lord.
}
\details{
The from_date and to_date arguments can be used to filter the speeches
returned. The on_date argument is a convenience that sets the from_date and
to_date to the same given date. The on_date has priority: if the on_date is
set, the from_date and to_date are ignored.

The filtering is inclusive: a speech is returned if the date of it falls
within the period specified with the from and to dates.
}
