% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_lord.R
\name{fetch_lords_maiden_speeches_raw}
\alias{fetch_lords_maiden_speeches_raw}
\title{Fetch maiden speeches: Lords}
\usage{
fetch_lords_maiden_speeches_raw()
}
\description{
Fetch maiden speeches: Lords
}
\keyword{internal}
