% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_lord.R
\name{fetch_lords_opposition_roles_raw}
\alias{fetch_lords_opposition_roles_raw}
\title{Fetch opposition roles: Lords}
\usage{
fetch_lords_opposition_roles_raw()
}
\description{
Fetch opposition roles: Lords
}
\keyword{internal}
