% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_lords.R
\name{fetch_lords_other_parliaments_raw}
\alias{fetch_lords_other_parliaments_raw}
\title{Fetch other parliaments: Lords}
\usage{
fetch_lords_other_parliaments_raw()
}
\description{
Fetch other parliaments: Lords
}
\keyword{internal}
