% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_lords.R
\name{fetch_lords_parliamentary_roles_raw}
\alias{fetch_lords_parliamentary_roles_raw}
\title{Fetch parliamentary roles: Lords}
\usage{
fetch_lords_parliamentary_roles_raw()
}
\description{
Fetch parliamentary roles: Lords
}
\keyword{internal}
