% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{fetch_members_blogs}
\alias{fetch_members_blogs}
\title{Fetch blogs for Members}
\usage{
fetch_members_blogs(
  fetch_members,
  fetch_addresses,
  address_func,
  from_date = NA,
  to_date = NA,
  on_date = NA
)
}
\description{
Fetch blogs for Members
}
\keyword{internal}
