% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{fetch_members_facebook}
\alias{fetch_members_facebook}
\title{Fetch Facebook accounts for Members}
\usage{
fetch_members_facebook(
  fetch_members,
  fetch_addresses,
  address_func,
  from_date = NA,
  to_date = NA,
  on_date = NA
)
}
\description{
Fetch Facebook accounts for Members
}
\keyword{internal}
