% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{fetch_members_office_addresses}
\alias{fetch_members_office_addresses}
\title{Fetch office addresses for Members}
\usage{
fetch_members_office_addresses(
  fetch_members,
  fetch_addresses,
  address_func,
  from_date = NA,
  to_date = NA,
  on_date = NA
)
}
\description{
Fetch office addresses for Members
}
\keyword{internal}
