% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{fetch_members_phone_numbers}
\alias{fetch_members_phone_numbers}
\title{Fetch phone numbers for Members}
\usage{
fetch_members_phone_numbers(
  fetch_members,
  fetch_addresses,
  address_func,
  from_date = NA,
  to_date = NA,
  on_date = NA
)
}
\description{
Fetch phone numbers for Members
}
\keyword{internal}
