% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps.R
\name{fetch_mps_addresses}
\alias{fetch_mps_addresses}
\title{Fetch addresses for all MPs}
\usage{
fetch_mps_addresses()
}
\value{
A tibble of addresses for each MP, with one row per address.
}
\description{
\code{fetch_mps_addresses} fetches data from the Members Names platform
showing contact details for each MP.
}
\details{
Addresses can represent contact information of different types, including
phsyical addresses, phone, fax, email, website, and social media. These
addresses are not time bound in MNIS so date filtering is not available for
this function.
}
