% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_mps.R
\name{fetch_mps_contested_elections_raw}
\alias{fetch_mps_contested_elections_raw}
\title{Fetch contested elections: MPs}
\usage{
fetch_mps_contested_elections_raw()
}
\description{
Fetch contested elections: MPs
}
\keyword{internal}
