% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts_mps.R
\name{fetch_mps_instagram}
\alias{fetch_mps_instagram}
\title{Fetch Instagram accounts for MPs}
\usage{
fetch_mps_instagram(from_date = NA, to_date = NA, on_date = NA)
}
\arguments{
\item{from_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the from_date.}

\item{to_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the to_date.}

\item{on_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the on_date.}
}
\value{
A tibble of known Instagram accounts for each MP, with one row per
  combination of MP and Instagram account.
}
\description{
\code{fetch_mps_instagram} fetches data from the Members Names platform on
Instagram accounts for each MP, with one row per combination of MP and
Instagram account.
}
\details{
The from_date and to_date arguments can be used to filter the MPs based
on the dates of their Commons memberships. The on_date argument is a
convenience that sets the from_date and to_date to the same given date. The
on_date has priority: if the on_date is set, the from_date and to_date are
ignored.

The filtering is inclusive: an MP is returned if any part of one of their
Commons memberships falls within the period specified with the from and to
dates.
}
