% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_mp.R
\name{fetch_mps_maiden_speeches_raw}
\alias{fetch_mps_maiden_speeches_raw}
\title{Fetch maiden speeches: MPs}
\usage{
fetch_mps_maiden_speeches_raw()
}
\description{
Fetch maiden speeches: MPs
}
\keyword{internal}
