% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_mp.R
\name{fetch_mps_opposition_roles_raw}
\alias{fetch_mps_opposition_roles_raw}
\title{Fetch opposition roles: MPs}
\usage{
fetch_mps_opposition_roles_raw()
}
\description{
Fetch opposition roles: MPs
}
\keyword{internal}
