% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_mps.R
\name{fetch_mps_other_parliaments_raw}
\alias{fetch_mps_other_parliaments_raw}
\title{Fetch other parliaments: MPs}
\usage{
fetch_mps_other_parliaments_raw()
}
\description{
Fetch other parliaments: MPs
}
\keyword{internal}
