% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fetch_query_data}
\alias{fetch_query_data}
\title{Fetch data from MNIS based on given query}
\usage{
fetch_query_data(house, data_output)
}
\description{
Fetch data from MNIS based on given query
}
\keyword{internal}
