% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{process_lords_output}
\alias{process_lords_output}
\title{Combine basic Lords data with output table}
\usage{
process_lords_output(output_table)
}
\description{
Combine basic Lords data with output table
}
\keyword{internal}
