% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{process_mps_output}
\alias{process_mps_output}
\title{Combine basic MP data with output table}
\usage{
process_mps_output(output_table)
}
\description{
Combine basic MP data with output table
}
\keyword{internal}
