% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clmnis.R
\docType{package}
\name{mnis}
\alias{mnis}
\title{clmnis: A package for downloading data from the Parliamentary Members Name
Information Service}
\description{
The clmnis package provides a suite of functions for downloading data from
the data platform for the UK Parliament.
}
